+++ test/PHPUnit/module/helios-generique/action/IParapheurEnvoieHeliosTest.php
@@ -23,7 +23,6 @@ class IParapheurEnvoieHeliosTest extends PastellTestCase
         $this->getObjectInstancier()->setInstance(SoapClientFactory::class, $soapClientFactory);
         $connector = $this->createConnector('iParapheur', 'parapheur');
         $this->configureConnector($connector['id_ce'], [
-            'iparapheur_activate' => true,
             'iparapheur_wsdl' => 'wsdl',
         ]);
         $this->associateFluxWithConnector($connector['id_ce'], 'helios-generique', 'signature');