+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 - Les classes spécifiques d'envoi en GED ne doivent plus être utilisées ou largement corrigées pour se baser sur connecteurt-type/GEDEnvoyer car il y a un risque de dépôts multiples si on attrape pas correctement les exceptions émisent par les connecteurs.
 - Les connecteurs suivants ont été retirés du cœur de Pastell pour être mis dans l'extension pastell-compat-v2 : ged-ftp, ged-ssh, ged-webdav, smb, cmis, recuperation-fichier-local, recuperation-fichier-ssh, glaneur-local, glaneur-doc, creation-pes-aller, creation-document (la plupart ne sont plus utilisables avec les flux du cœur Pastell) #672
 - Les inscriptions "citoyen" et "fournisseur" sont retirées car non-utilisées
+- Suppression de la case à cocher "Module activé" dans le connecteur i-Parapheur
 
 ## Dépréciations